requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
groq
